RotationSettings.py

+++ src/mob/SessionSettings/RotationSettings.py
@@ -6,6 +6,6 @@ from dataclasses_json import dataclass_json
 @dataclass_json
 @dataclass(frozen=True)
 class RotationSettings:
-    driverInMinutes: int = 10
-    breakInMinutes: int = 15
+    driverInMinutes: float = 10
+    breakInMinutes: float = 15
     howManyRotationsBeforeBreak: int = 6
